refactor(vscode): extract workspace runtime state (#4255) (#4384)
