+++ vscode-extension/scripts/vsix-inventory-baseline.json
@@ -1,7 +1,7 @@
 {
   "schema_version": 1,
   "total_files": 28,
-  "total_bytes": 1472504,
+  "total_bytes": 1472562,
   "files": {
     "README.md": 21701,
     "package.json": 53908,
@@ -12,11 +12,11 @@
     "CHANGELOG.md": 16170,
     ".oxlintrc.json": 559,
     ".oxfmtrc.json": 203,
-    "syntaxes/perl.tmLanguage.json": 11634,
-    "syntaxes/gherkin.tmLanguage.json": 1875,
     "snippets/perl.json": 12172,
     "snippets/launch.json": 1139,
-    "out/extension.js": 1314887,
+    "syntaxes/perl.tmLanguage.json": 11634,
+    "syntaxes/gherkin.tmLanguage.json": 1875,
+    "out/extension.js": 1314945,
     "media/walkthrough/welcome.svg": 756,
     "media/walkthrough/verify-perl.svg": 1189,
     "media/walkthrough/try-goto-definition.svg": 1852,
